megaphone: split markets and tickers into separate files.
It makes sense to split tickers and markets into separate classes
due to classes size to make it easily readable and editable.
